pyproject.toml
[project]
name = "electric-demand-data"
version = "0.1.0"
description = "Electricity demand data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cdsapi>=0.7.5",
    "countryinfo>=0.1.2",
    "entsoe-py>=0.6.18",
    "geopandas>=1.0.1",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.0",
    "mkdocs-material>=9.5.50",
    "pandas>=2.2.3",
    "pathlib>=1.0.1",
    "pyarrow>=19.0.0",
    "pycountry>=24.6.1",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "timezonefinder>=6.5.8",
    "xarray>=2025.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
